requirements.txt
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
python-decouple==3.8
mysqlclient==2.2.4
Pillow==10.4.0
drf-spectacular==0.27.0